Move `ChatSection` to its own directory
